change "compiled on server..." to "run"
I think a good distinction here is that web code is rarely (never?) compiled in either environment, per se. Compiled code like Java, Go, etc are compiled beforehand and then deployed to the server. The code will be **run** on either machine / environment and "executed" or "interpreted" is a better word for what happens but I thought this wording might be more accurate / not introduce a vocab word "compile" that could be confusing when learning about compiled languages later (if ever).